Add support get claims and to fetch userinfo
The API now returns the standard and extra claims of the validated token
together with the a type indicator if that token is a special token
(access, refresh as known by kc).

With a valid access token, the userinfo can now be fetched using a new
method.

The validate examples have been updated accordingly. The
userinfo is returned as UTF-8 encoded JSON string.