Use backend identifier claim for ID in validation
This is a change in behavior. Whenever a access token is validated with
contains claims from Kopano Konnect, specially the kc.i.id subclaim in
kc.identity, the validation will now return that value instead of the
subject. Implementations can still fetch the token subject claim by
fetching it directly from the standard claims result. If the token does
not contain the kc.i.id claim, then the subject value will be used like
it was before.